feat: [M3-7114] Break large component into smaller ones
